Modified header, pointer to unsigned long, process char, process poin…

+++ holberton.h
@@ -34,7 +34,7 @@ void rev_string(char *s);
 /* util2.c */
 void dec_converter(long dec, char *str_converted, int base);
 
-void dec_converter_uns(unsigned int dec, char *str_converted, int base);
+void dec_converter_uns(unsigned long dec, char *str_converted, int base);
 
 void translate_format(char *buffer, char *format_buffer);
